+++ ark/type/__tests__/integration/config.ts
@@ -1,0 +1,3 @@
+import { configure } from "arktype/config"
+
+configure({ numberAllowsNaN: true })