+++ ark/type/__tests__/integration/importConfig.ts
@@ -1,7 +1,8 @@
-import { configure } from "arktype/config"
-configure({ numberAllowsNaN: true })
+import "./config.ts"
 
 import { type } from "arktype"
 
 if (!type("number").allows(Number.NaN))
 	throw new Error(`global configuration not applied`)
+
+console.log("✅ importConfig passed")